+++ tests/test_middleware.py
@@ -23,11 +23,14 @@ async def test_500(cli, caplog):
     assert 'custom 500 error' == await r.text()
     assert len(caplog.records) == 1
     record = caplog.records[0]
-    assert record.data.keys() == {'request_duration', 'request', 'response'}
-    assert record.data['request']['text'] == 'foobar'
+    assert record.data.keys() == {'request_duration', 'response'}
+    assert record.request['url'].startswith('http://127.0.0.1:')
+    assert record.request['data'] == 'foobar'
+    assert record.request['method'] == 'GET'
+    assert record.request['cookies'] == []
+    assert ('Accept', '*/*') in record.request['headers']
     assert record.data['response']['text'] == 'custom 500 error'
     assert record.user == {'ip_address': '127.0.0.1', 'username': 'foobar'}
-    assert record.tags == {}
     assert record.fingerprint == ('/errors/{do}', '500')
 
 
@@ -37,10 +40,9 @@ async def test_503_with_name(cli, caplog):
     assert 'test response with status 503' == await r.text()
     assert len(caplog.records) == 1
     record = caplog.records[0]
-    assert record.data.keys() == {'request_duration', 'request', 'response'}
+    assert record.data.keys() == {'request_duration', 'response'}
     assert record.data['response']['text'] == 'test response with status 503'
     assert record.user == {'ip_address': '127.0.0.1', 'username': 'foobar'}
-    assert record.tags == {}
     assert record.fingerprint == ('any-status', '503')
 
 
@@ -50,11 +52,10 @@ async def test_405(cli, caplog):
     assert '405: Method Not Allowed' == await r.text()
     assert len(caplog.records) == 1
     record = caplog.records[0]
-    assert record.data.keys() == {'request_duration', 'request', 'response'}
-    assert record.data['request']['text'] == ''
+    assert record.data.keys() == {'request_duration', 'response'}
+    assert record.request['data'] == ''
     assert record.data['response']['text'] == '405: Method Not Allowed'
     assert record.user == {'ip_address': '127.0.0.1', 'username': 'foobar'}
-    assert record.tags == {}
     assert record.fingerprint == ('/', '405')
 
 
@@ -64,10 +65,9 @@ async def test_not_unicode(cli, caplog):
     assert 'custom 500 error' == await r.text()
     assert len(caplog.records) == 1
     record = caplog.records[0]
-    assert record.data.keys() == {'request_duration', 'request', 'response'}
-    assert record.data['request']['text'] is None
+    assert record.data.keys() == {'request_duration', 'response'}
+    assert record.request['data'] is None
     assert record.user == {'ip_address': '127.0.0.1', 'username': 'foobar'}
-    assert record.tags == {}
 
 
 async def test_499(cli, caplog):
@@ -75,9 +75,8 @@ async def test_499(cli, caplog):
     assert r.status == 499, await r.text()
     assert len(caplog.records) == 2
     record = caplog.records[1]
-    assert record.data.keys() == {'request_duration', 'request', 'response'}
+    assert record.data.keys() == {'request_duration', 'response'}
     assert record.user == {'ip_address': '127.0.0.1'}
-    assert record.tags == {}
 
 
 async def test_value_error(cli, caplog):
@@ -86,10 +85,9 @@ async def test_value_error(cli, caplog):
     assert '500: Internal Server Error' == await r.text()
     assert len(caplog.records) == 1
     record = caplog.records[0]
-    assert record.data.keys() == {'request_duration', 'request', 'response', 'exception_extra'}
+    assert record.data.keys() == {'request_duration', 'response', 'exception_extra'}
     assert record.data['exception_extra'] is None
     assert record.user == {'ip_address': '127.0.0.1', 'username': 'foobar'}
-    assert record.tags == {}
 
 
 async def test_user(cli, caplog):
@@ -97,9 +95,8 @@ async def test_user(cli, caplog):
     assert r.status == 488, await r.text()
     assert len(caplog.records) == 1
     record = caplog.records[0]
-    assert record.data.keys() == {'request_duration', 'request', 'response'}
+    assert record.data.keys() == {'request_duration', 'response'}
     assert record.user == {'ip_address': '127.0.0.1', 'username': 'foobar'}
-    assert record.tags == {}
 
 
 def test_exc_extra_ok():